+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Added
+
+- Permission check. Only show the widget if the user has permissions to view the Member Audit app
+
 ## \[1.0.6\] - 2024-09-22 ([@ppfeufer](https://github.com/ppfeufer))
 
 ### Added